+++ test/Consumer/namedConsumer.test.tsx
@@ -98,6 +98,101 @@ it('throw when not provided', async () => {
   jest.restoreAllMocks()
 })
 
+const UndefComsumerComponent = () => {
+  return (
+    <>
+      <Consumer<number> allowUndef name='test-number'>
+        {(number) => <h2 data-testid='number'>{number}</h2>}
+      </Consumer>
+      <Consumer<boolean> allowUndef name='test-boolean'>
+        {(boolean) => <h2 data-testid='boolean'>{boolean ? 'true' : 'false'}</h2>}
+      </Consumer>
+      <Consumer<string> allowUndef name='test-text'>
+        {(text) => <h2 data-testid='text'>{text}</h2>}
+      </Consumer>
+      <Consumer<Counter> allowUndef name='test-counter'>
+        {(counter) => <h2 data-testid='count'>{counter?.count}</h2>}
+      </Consumer>
+      <Consumer<CustomData> allowUndef name='test-custom'>
+        {(customData) => <h2 data-testid='key'>{customData?.key}</h2>}
+      </Consumer>
+      <Consumer<CustomData> allowUndef name='test-custom'>
+        {(customData) => <h2 data-testid='value'>{customData?.value}</h2>}
+      </Consumer>
+    </>
+  )
+}
+
+it('not throw when not provided and allowUndef', () => {
+  const { container } = render(<UndefComsumerComponent />)
+  const renderedNumberElement = getByTestId(container, 'number')
+  const renderedNumber = renderedNumberElement.textContent
+  const expectedNumber = ''
+  expect(renderedNumber).toBe(expectedNumber)
+
+  const renderedBooleanElement = getByTestId(container, 'boolean')
+  const renderedBoolean = renderedBooleanElement.textContent
+  const expectedBoolean = 'false'
+  expect(renderedBoolean).toBe(expectedBoolean)
+
+  const renderedTextElement = getByTestId(container, 'text')
+  const renderedText = renderedTextElement.textContent
+  const expectedText = ''
+  expect(renderedText).toBe(expectedText)
+
+  const renderedCountElement = getByTestId(container, 'count')
+  const renderedCount = renderedCountElement.textContent
+  const expectedCount = ''
+  expect(renderedCount).toBe(expectedCount)
+
+  const renderedKeyElement = getByTestId(container, 'key')
+  const renderedKey = renderedKeyElement.textContent
+  const expectedKey = ''
+  expect(renderedKey).toBe(expectedKey)
+
+  const renderedValueElement = getByTestId(container, 'value')
+  const renderedValue = renderedValueElement.textContent
+  const expectedValue = ''
+  expect(renderedValue).toBe(expectedValue)
+})
+
+it('not throw when provided and allowUndef', () => {
+  const { container } = render(
+    <ProviderComponent>
+      <UndefComsumerComponent />
+    </ProviderComponent>,
+  )
+  const renderedNumberElement = getByTestId(container, 'number')
+  const renderedNumber = renderedNumberElement.textContent
+  const expectedNumber = '5'
+  expect(renderedNumber).toBe(expectedNumber)
+
+  const renderedBooleanElement = getByTestId(container, 'boolean')
+  const renderedBoolean = renderedBooleanElement.textContent
+  const expectedBoolean = 'true'
+  expect(renderedBoolean).toBe(expectedBoolean)
+
+  const renderedTextElement = getByTestId(container, 'text')
+  const renderedText = renderedTextElement.textContent
+  const expectedText = 'test-text'
+  expect(renderedText).toBe(expectedText)
+
+  const renderedCountElement = getByTestId(container, 'count')
+  const renderedCount = renderedCountElement.textContent
+  const expectedCount = '6'
+  expect(renderedCount).toBe(expectedCount)
+
+  const renderedKeyElement = getByTestId(container, 'key')
+  const renderedKey = renderedKeyElement.textContent
+  const expectedKey = 'test-key'
+  expect(renderedKey).toBe(expectedKey)
+
+  const renderedValueElement = getByTestId(container, 'value')
+  const renderedValue = renderedValueElement.textContent
+  const expectedValue = 'test-val'
+  expect(renderedValue).toBe(expectedValue)
+})
+
 const ProviderComponent2 = ({ children }: PropsWithChildren) => {
   const [count, setCount] = useState(0)
   return (
@@ -109,7 +204,6 @@ const ProviderComponent2 = ({ children }: PropsWithChildren) => {
     </Provider>
   )
 }
-
 const ComsumerComponent2 = () => {
   const number = useNamedProvider<number>('test-count')
   return <h2 data-testid='number'>{number}</h2>